visa is okay

+++ src/main/java/com/contractEmployee/contractEmployee/rep/EmployeeRepository.java
@@ -2,12 +2,12 @@ package com.contractEmployee.contractEmployee.rep;
 
 import com.contractEmployee.contractEmployee.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.contractEmployee.contractEmployee.entity.Contract;
+//import com.contractEmployee.contractEmployee.entity.Contract;
 
 
 import java.util.List;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
     boolean existsByStaffCode(String staffCode);
 
 }
